Fix to display results with no labels
The dotplot command will not attempt to generate labels when the label
flag is passed. The job will not expect labels to be returned as a
result when the clabel attribute is zero. The results window disables
the x and y labels and will only warn when the labels where expected to
be displayed.

* Add check for labels flag in dotplot
* Labels are inputs when clabel is 1
* Show labels when clabel is 1

	modified:   web/SynMap_new.pl
	modified:   web/bin/dotplot.pl
	modified:   web/tmpl/partials/synmap_results.tmpl