add basic analysis tracking to analytics
Analytics have been added for running an analysis. The tool records an
event for click to run the analysis. The event category is based on the
tool with the action being `run` and the label/value left blank with
the exception of CoGeBlast which specifies if the tool ran ncbi's or
coge's blast.

	modified:   tmpl/CoGeBlast.tmpl
	modified:   tmpl/CoGeBlast_ncbi.tmpl
	modified:   tmpl/GEvo.tmpl
	modified:   tmpl/SynFind.tmpl
	modified:   tmpl/SynMap.tmpl