disable layers that are too large
Layers too large to display in the browser are not shown. The layer is
disabled and a title is set on its label. The previous set of layers
has been split into seperate files. The plot loads the syntenic data
and will load the non-syntenic data if the data isn't larger than
268435456 bytes.

	modified:   SynMap_new.pl
	modified:   bin/dotplot_dots.pl
	modified:   js/synmap.js
	modified:   tmpl/partials/synmap_results.tmpl