add support for posting json content
A request can now set the content type as application/json and the
POSTDATA will be parsed as a json string. If the data is not valid json
a PAYLOAD error is returned. If the PAYLOAD can not be properly routed
a NOT_FOUND error is returned. When successful the json object will
submitted as the arguments to the dispatch function.

	modified:   modules/Accessory/lib/CoGe/Accessory/Web.pm